requirements.txt
Flask==3.0.3
Pillow==10.3.0
google-generativeai==0.7.1
